April 2023 updates
Code adjusted to remove D-Wave from Amazon Braket SDK. Install D-Wave using pip install dwave-ocean-sdk && dwave setup. Aspen 11, M-1 and M-2 are replaced by Aspen-M-3. Some code is not re-run or modified to keep the results in line with the book. In other places, fixes have been inserted, so that the code should execute properly with D-Wave and Aspen changes. QuEra Aquila was also added to Amazon Braket, however, it is not used in the code. Chapter 11 and 12 have the original code from 2022 for reference and matches the results and figures in the book.